# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ODIN Swagger UI",
        "title": "Odin: Remote Administration for 1S.Enterprise Application Servers",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "khorevaa@yandex.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/api/v1",
    "paths": {
...
        "/app/{app}/agent/admins": {
            "get": {
                "description": "получение списка администраторов агента на сервере 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admins",
                    "agent"
                ],
                "summary": "получение списка администраторов агента на сервере 1С Предприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "force update ignore cache",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/serialize.UserInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            },
            "post": {
                "description": "выполняет регистрацию нового адмнимистратор на агенте сервера 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admins",
                    "agent"
                ],
                "summary": "выполняет регистрацию нового адмнимистратор на агенте сервера 1С Предприятиеи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user info",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serialize.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/serialize.UserInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
        "/app/{app}/agent/admins/{admin}": {
            "delete": {
                "description": "Удаление администратора агента на сервере 1С Предприятие",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admins",
                    "agent"
                ],
                "summary": "Удаление администратора агента на сервере 1С Предприятие",
                "parameters": [
                    {
                        "type": "string",
                        "description": "app name",
                        "name": "app",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "admin name",
                        "name": "admin",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "force update ignore cache",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "serialize.UserInfo": {
            "type": "object",
            "properties": {
                "descr": {
                    "type": "string",
                    "example": "дополнительное описание"
                },
                "name": {
                    "type": "string",
                    "example": "admin"
                },
                "password": {
                    "type": "string"
                },
                "password_auth_allowed": {
                    "type": "boolean"
                },
                "sys_auth_allowed": {
                    "type": "boolean"
                },
                "sys_user_name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "metrics": " Grants read access and control metrics operations",
                "read": " Grants read access",
                "write": " Grants write access"
            }
        }
    }
}